+++ tests/test_knapsackStrategies.py
@@ -3,7 +3,7 @@ from click.testing import CliRunner
 from typing import List
 from algorithm_tester.tester_logic import test_instance_file
 from algorithm_tester.tested_algorithms import Algorithms
-from algorithm_tester.mydataclasses import Solution
+from algorithm_tester.tester_dataclasses import Solution
 from algorithm_tester.helpers import FilePair, getFiles
 
 @pytest.mark.parametrize(
